Fixed bug#2626 - Production - Exception thrown when claim number contains an apostrophe (').
	modified:   idas.chox.core/src/main/java/idas/chox/core/services/ClaimService.java
	modified:   idas.chox.data/src/main/java/idas/chox/data/services/BaseDataService.java
	modified:   idas.chox.data/src/main/java/idas/chox/data/services/ClaimServiceImpl.java

+++ idas.chox.core/src/main/java/idas/chox/core/services/ClaimService.java
@@ -12,11 +12,11 @@ public interface ClaimService extends DataService {
 
     Claim getClaim(int id);
 
-    Long getClaimCountByClaimNumber(String claimNumber, int claimId);
+    int getClaimCountByClaimNumber(String claimNumber, int claimId);
 
     List getOtherClaimsByClaimNumber(String claimNumber, int claimId);
 
-    Long getECDCountByClaimId(int claimId);
+    int getECDCountByClaimId(int claimId);
 
     Integer getCountOfClaimByVRN(String strVRN, int claimId);
